+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog],
 and this project adheres to [Semantic Versioning].
 
+## [Unreleased]
+
+### Added
+- Export to XLSX functionality - export the mod list to an Excel file for easy sharing and backup.
+
 ## [1.2.0] - 2025-09-18
 
 ### Added
@@ -70,6 +75,7 @@ and this project adheres to [Semantic Versioning].
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
 <!-- Versions -->
+[unreleased]:  https://github.com/VidyakinIvan/Modding-Assistant/compare/Modding-Assistant%2Fv1.2.0...HEAD
 [1.2.0]: https://github.com/VidyakinIvan/Modding-Assistant/compare/Modding-Assistant%2Fv1.1.0...Modding-Assistant%2Fv1.2.0
 [1.1.0]: https://github.com/VidyakinIvan/Modding-Assistant/compare/Modding-Assistant%2Fv1.0.0...Modding-Assistant%2Fv1.1.0
 [1.0.0]: https://github.com/VidyakinIvan/Modding-Assistant/releases/tag/Modding-Assistant%2Fv1.0.0